refactor: 테스트 코드 닉네임 제거
